review comments

+++ src/LinuxTracing/PerfEventOpen.cpp
@@ -70,20 +70,28 @@ int mmap_task_event_open(pid_t pid, int32_t cpu) {
   return generic_event_open(&pe, pid, cpu);
 }
 
-int stack_sample_event_open(uint64_t period_ns, pid_t pid, int32_t cpu, uint32_t stack_dump_size) {
+int stack_sample_event_open(uint64_t period_ns, pid_t pid, int32_t cpu, uint16_t stack_dump_size) {
   perf_event_attr pe = generic_event_attr();
   pe.type = PERF_TYPE_SOFTWARE;
   pe.config = PERF_COUNT_SW_CPU_CLOCK;
   pe.sample_period = period_ns;
   pe.sample_type |= PERF_SAMPLE_REGS_USER | PERF_SAMPLE_STACK_USER;
   pe.sample_regs_user = SAMPLE_REGS_USER_ALL;
+
+  // Max to pass to perf_event_open without getting an error is (1u << 16u) - 8,
+  // because the kernel stores this in a short and because of alignment reasons.
+  // But the size the kernel actually returns is smaller, because the maximum size
+  // of the entire record the kernel is willing to return is (1u << 16u) - 8.
+  // If we want the size we pass to coincide with the size we get, we need to pass
+  // a lower value. For the current layout of perf_event_stack_sample_fixed, the maximum
+  // size is 65312. We leave some extra room with our flag (see `ClientFlags.cpp`).
   pe.sample_stack_user = stack_dump_size;
 
   return generic_event_open(&pe, pid, cpu);
 }
 
 int callchain_sample_event_open(uint64_t period_ns, pid_t pid, int32_t cpu,
-                                uint32_t stack_dump_size) {
+                                uint16_t stack_dump_size) {
   perf_event_attr pe = generic_event_attr();
   pe.type = PERF_TYPE_SOFTWARE;
   pe.config = PERF_COUNT_SW_CPU_CLOCK;
