+++ src/OrbitGl/ClientFlags.cpp
@@ -29,12 +29,9 @@ ABSL_FLAG(uint16_t, sampling_rate, 1000, "Frequency of callstack sampling in sam
 
 // Max to pass to perf_event_open without getting an error is (1u << 16u) - 8,
 // because the kernel stores this in a short and because of alignment reasons.
-// But the size the kernel actually returns is smaller, because the maximum size
-// of the entire record the kernel is willing to return is (1u << 16u) - 8.
-// If we want the size we pass to coincide with the size we get, we need to pass
-// a lower value. For the current layout of perf_event_stack_sample_fixed, the maximum
-// size is 65312, but let's leave some extra room.
-ABSL_FLAG(uint32_t, stack_dump_size, 65000,
+// But the size the kernel actually returns is smaller and we leave some extra room (see
+// PerfEventOpen.cpp`).
+ABSL_FLAG(uint16_t, stack_dump_size, 65000,
           "Number of bytes to copy from the stack per sample. Max: 65000");
 
 // TODO(b/160549506): Remove this flag once it can be specified in the ui.
